Simplify Resolution API
