+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/DebianResolutionApi.java
@@ -18,9 +18,16 @@
 
 package eu.fasten.analyzer.restapiplugin.mvn.api;
 
+import eu.fasten.analyzer.restapiplugin.mvn.KnowledgeBaseConnector;
+import eu.fasten.core.dependents.GraphResolver;
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -31,10 +38,18 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class DebianResolutionApi {
 
-    private final DebianResolutionApiService service;
+    private static final Logger logger = LoggerFactory.getLogger(DebianResolutionApi.class);
+    private GraphResolver graphDebianResolver;
 
-    public DebianResolutionApi(DebianResolutionApiService service) {
-        this.service = service;
+    public DebianResolutionApi() {
+        try {
+            var graphDebianResolver = new GraphResolver();
+            graphDebianResolver.buildDependencyGraph(KnowledgeBaseConnector.dbCContext, KnowledgeBaseConnector.dependencyDebianGraphPath);
+            this.graphDebianResolver = graphDebianResolver;
+        } catch (Exception e) {
+            logger.error("Error constructing Debian dependency graph resolver", e);
+            System.exit(1);
+        }
     }
 
     @GetMapping(value = "/debian/packages/{pkg}/{pkg_ver}/resolve/dependents", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -42,6 +57,19 @@ public class DebianResolutionApi {
                                              @PathVariable("pkg_ver") String package_version,
                                              @RequestParam(required = false, defaultValue = "true") boolean transitive,
                                              @RequestParam(required = false, defaultValue = "-1") long timestamp) {
-        return service.resolveDependents(package_name, package_version, transitive, timestamp);
+        timestamp = ((timestamp == -1) ? this.graphDebianResolver.getCreatedAt(package_name,package_version, KnowledgeBaseConnector.dbCContext): timestamp);
+        var depSet = this.graphDebianResolver.resolveDependents(package_name,
+           package_version, timestamp, transitive);
+        var jsonArray = new JSONArray();
+        depSet.stream().map(eu.fasten.core.dependents.data.Revision::toJSON).peek(json -> {
+            var dep_name = json.getString("package");
+            var ver = json.getString("version");
+            var url = String.format("%sdebian/callgraphs/%s/%s/buster/%s/amd64/file.json", KnowledgeBaseConnector.rcgBaseUrl,
+                    dep_name.charAt(0), dep_name, ver);
+            json.put("url", url);
+        }).forEach(jsonArray::put);
+        var result = jsonArray.toString();
+        result = result.replace("\\/", "/");
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 }
