+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/PyPiResolutionApi.java
@@ -18,6 +18,14 @@
 
 package eu.fasten.analyzer.restapiplugin.mvn.api;
 
+import eu.fasten.analyzer.restapiplugin.mvn.KnowledgeBaseConnector;
+import eu.fasten.core.dependents.GraphResolver;
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
+
 import org.springframework.context.annotation.Lazy;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -31,10 +39,18 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class PyPiResolutionApi {
 
-    private final PyPiResolutionApiService service;
+    private static final Logger logger = LoggerFactory.getLogger(PyPiResolutionApi.class);
+    private GraphResolver graphPyPiResolver;
 
-    public PyPiResolutionApi(PyPiResolutionApiService service) {
-        this.service = service;
+    public PyPiResolutionApi() {
+        try {
+            var graphPyPiResolver = new GraphResolver();
+            graphPyPiResolver.buildDependencyGraph(KnowledgeBaseConnector.dbCContext, KnowledgeBaseConnector.dependencyPyPiGraphPath);
+            this.graphPyPiResolver = graphPyPiResolver;
+        } catch (Exception e) {
+            logger.error("Error constructing PyPi dependency graph resolver", e);
+            System.exit(1);
+        }
     }
 
     @GetMapping(value = "/pypi/packages/{pkg}/{pkg_ver}/resolve/dependents", produces = MediaType.APPLICATION_JSON_VALUE)
@@ -42,6 +58,20 @@ public class PyPiResolutionApi {
                                              @PathVariable("pkg_ver") String package_version,
                                              @RequestParam(required = false, defaultValue = "true") boolean transitive,
                                              @RequestParam(required = false, defaultValue = "-1") long timestamp) {
-        return service.resolveDependents(package_name, package_version, transitive, timestamp);
+
+        timestamp = ((timestamp == -1) ? this.graphPyPiResolver.getCreatedAt(package_name, package_version, KnowledgeBaseConnector.dbPythonContext): timestamp);
+        var depSet = this.graphPyPiResolver.resolveDependents(package_name,
+            package_version, timestamp, transitive);
+        var jsonArray = new JSONArray();
+        depSet.stream().map(eu.fasten.core.dependents.data.Revision::toJSON).peek(json -> {
+            var dep_name = json.getString("package");
+            var ver = json.getString("version");
+            var url = String.format("%spypi/pypi/callgraphs/%s/%s/%s/cg.json", KnowledgeBaseConnector.rcgBaseUrl,
+                    dep_name.charAt(0), dep_name, ver);
+            json.put("url", url);
+        }).forEach(jsonArray::put);
+        var result = jsonArray.toString();
+        result = result.replace("\\/", "/");
+        return new ResponseEntity<>(result, HttpStatus.OK);
     }
 }
